+++ pkg/shell/syft.go
@@ -16,9 +16,9 @@ func (s *syftCmd) Version() *Command {
 
 // Run executes the Syft CLI
 //
-// shell: `syft docker-archive:${IMAGE_TARBALL} --scope=squashed -o cyclonedx-json=${IMAGE_SBOM}`
-func (s *syftCmd) ScanImage(imageTarball string, imageSbom string) *Command {
-	cmd := s.InitCmd().WithArgs("docker-archive:"+imageTarball, "--scope=squashed", "-o", "cyclonedx-json="+imageSbom)
+// shell: `syft <image> --scope=squashed -o cyclonedx-json`
+func (s *syftCmd) ScanImage(image string, sbomFilename string) *Command {
+	cmd := s.InitCmd().WithArgs(image, "--scope=squashed", "-o", "cyclonedx-json")
 
 	return NewCommand(cmd)
 }